==> data_analysis/registered_report/analysis/manuscript/3_compute_RT.Rmd <==
---
title: "CATegories Computing RTs"
author: "Martin Zettersten"
date: "2024-03-12"
output: html_document
---

## Load packages

```{r setup, message=FALSE}
library(tidyverse)
library(here)
source("rt_helper.R")
source("resampling_helper.R")

knitr::opts_chunk$set(cache = FALSE, warn = FALSE,warning=FALSE, message = FALSE)
```

## Load data

```{r global_params}
data_file_path <- here::here("..","..","data","processed_data","CATegories_exp2_processed_data_with_exclusion_info.csv")
d <- read_csv(data_file_path)
```

## Resample times

We resample times here at a slightly higher sampling rate than in the subsequent cluster permutation analysis, in order to preserve more fine-grained information about timing for RT analyses.

```{r message=FALSE, warning=FALSE, results=FALSE}
#resample times
resampled_d <- d %>%
  mutate(
    aoi = case_when(
      accuracy_transformed == 1 ~ "target",
      accuracy_transformed == 0 ~ "distractor",
      is.na(accuracy_transformed) ~ "other"
    )
  ) %>%
  mutate(t_norm=time_normalized_corrected,
         trial_id = trial_number,
         administration_id = paste(sub_num,session,sep="_")) %>%
  resample_times(sample_duration=1000/30) #time resampling is set to 30 Hz, to err on the side of preserving timing information
```

## Convert to run-length encoding format

This helps prepare the data for computing RTs

```{r}
#convert to rle data
rle_data <- resampled_d %>%
  filter(any(t_norm == 0), # must have data at 0
         t_norm >= 0) %>% # only pass data after 0 
  separate(administration_id,into=c("exp_id","sub_number","session"),sep="_",remove=FALSE) %>%
  unite("sub_num",exp_id,sub_number,sep="_") %>%
  mutate(trial_number=trial_id) %>%
  group_by(sub_num, session,trial_number) %>%
  summarise(lengths = rle(aoi)$lengths, 
            values = rle(aoi)$values)
```

## Compute RTs

```{r}
# compute RTs
d_rt <- rle_data %>%
  group_by(sub_num, session,trial_number) %>%
  nest() %>%
  mutate(data = lapply(data, get_rt)) %>%
  unnest(cols = c(data))
```

## Join in key exclusion info

```{r}
#join in key exclusion info
exclusion_info <- d %>% 
  distinct(sub_num,session,trial_number,trial_exclusion,trial_exclusion_reason,exclude_participant)
d_rt <- d_rt %>%
  mutate(session=as.numeric(session)) %>%
  left_join(exclusion_info)
```

## Write data

```{r}
# write RT data frame
write_csv(d_rt,here::here("..","..","data","processed_data","CATegories_exp2_RT_by_trial.csv"))
```

## Visually check data

```{r}
#quick plot of distribution
hist(filter(d_rt, shift_type=="D-T")$shift_start_rt)
```
